Drastically simplify LSP tests
