+++ crates/lsp/src/test.rs
@@ -2,6 +2,6 @@ mod client;
 mod client_ext;
 mod utils;
 
-pub(crate) use client::with_client;
+pub(crate) use client::new_test_client;
 pub(crate) use client_ext::TestClientExt;
 pub(crate) use utils::extract_marked_range;
